Simplify template implementation
