+++ crates/summit/Cargo.toml
@@ -13,7 +13,6 @@ service = { path = "../service" }
 
 async-trait.workspace = true
 axum.workspace = true
-camino = { workspace = true, optional = true }
 chrono.workspace = true
 clap.workspace = true
 color-eyre.workspace = true
@@ -23,10 +22,8 @@ futures-util.workspace = true
 hex.workspace = true
 http.workspace = true
 http-serde.workspace = true
-include_dir = { workspace = true, optional = true }
 itertools.workspace = true
 minijinja.workspace = true
-minijinja-autoreload = { workspace = true, optional = true }
 moss.workspace = true
 serde.workspace = true
 sha2.workspace = true
@@ -42,3 +39,7 @@ tower.workspace = true
 tower-http.workspace = true
 tracing.workspace = true
 uuid.workspace = true
+
+camino = { workspace = true, optional = true }
+include_dir = { workspace = true, optional = true }
+minijinja-autoreload = { workspace = true, optional = true }
